fix minor errors
