+++ cartesian_vic_controller/config/example_admittance_controller_config.yaml
@@ -37,7 +37,6 @@ cartesian_vic_controller:
     external_torque_sensor:
       is_enabled: false
       name: ""
-      filter_coefficient: 0.1 # 1 - exp(-Ts/tau) = 1 - exp(-Ts*2*pi*f_cutoff)
 
     control:
       frame:
@@ -90,9 +89,9 @@ cartesian_vic_controller:
         - 0.9 # x
         - 0.9 # y
         - 0.9 # z
-        - 0.99 # rx
-        - 0.99 # ry
-        - 0.99 # rz
+        - 0.9 # rx
+        - 0.9 # ry
+        - 0.9 # rz
 
       stiffness:
         - 200.0
