+++ cartesian_vic_controller/src/utils.cpp
@@ -16,7 +16,7 @@
 
 #include <cartesian_vic_controller/utils.hpp>
 
-namespace cartesian_admittance_controller
+namespace cartesian_vic_controller
 {
 
 geometry_msgs::msg::Accel AccelToMsg(const Eigen::Matrix<double, 6, 1> & in)
@@ -41,27 +41,6 @@ geometry_msgs::msg::Wrench WrenchToMsg(const Eigen::Matrix<double, 6, 1> & in)
   msg.torque.y = in[4];
   msg.torque.z = in[5];
   return msg;
-}
-
-template<class Derived>
-void matrixEigenToMsg(const Eigen::MatrixBase<Derived> & e, std_msgs::msg::Float64MultiArray & m)
-{
-  if (m.layout.dim.size() != 2) {
-    m.layout.dim.resize(2);
-  }
-  m.layout.dim[0].stride = e.rows() * e.cols();
-  m.layout.dim[0].size = e.rows();
-  m.layout.dim[1].stride = e.cols();
-  m.layout.dim[1].size = e.cols();
-  if (static_cast<int>(m.data.size()) != e.size()) {
-    m.data.resize(e.size());
-  }
-  int ii = 0;
-  for (int i = 0; i < e.rows(); ++i) {
-    for (int j = 0; j < e.cols(); ++j) {
-      m.data[ii++] = e.coeff(i, j);
-    }
-  }
 }
 
 bool fromMsg(const std_msgs::msg::Float64MultiArray & m, Eigen::Matrix<double, 6, 6> & e)
@@ -95,4 +74,4 @@ unsigned int flattened_index_from_triangular_index(unsigned int idx_row, unsigne
   return i * (i - 1) / 2 + j;
 }
 
-}  // namespace cartesian_admittance_controller
+}  // namespace cartesian_vic_controller
